add header install; reorganize prng files
