+++ aes/CMakeLists.txt
@@ -21,4 +21,5 @@
 cmake_minimum_required(VERSION 3.0)
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wall -Wextra -O3 -maes -mssse3 -msse4.1")
 add_library(flo-aesni STATIC flo-aesni.h aesni.c)
-install(TARGETS flo-aesni DESTINATION lib)
+set_target_properties(flo-aesni PROPERTIES PUBLIC_HEADER "flo-aesni.h")
+install(TARGETS flo-aesni LIBRARY DESTINATION lib PUBLIC_HEADER DESTINATION include)
